bin/tosrec: Switch to avconv. Auto make mp3 at end. Workaround -t bug.
Replace ffmpeg with avconv because the former is deprecated.
avconv has a bug where, if -t is specified, no sound will be recorded
(It will say that sound is being recorded, but the output file actually
lacks an audio stream!).

When -t is given, automatically run tomp3 at the end to make an audio-only
copy of the recording.