notes/12.04-want-packes: Created
This is a list of all the packages that I currently want to have installed
on a Ubuntu 12.04 "Precise" machine and that can be found in the stock
repositories.

Found by comparing the output of "apt-mark showmanual" from a clean install
with the same from a well-configured machine, followed by manual editing
out of specific kernel version and hardware-specific items (e.g. graphics
drivers).

Install all of these to quickly bring a fresh installation mostly up to
speed.

Does not include things from PPA's.