tmux: Better window label colors: Yellow for finished, Red for error
The "current" window is now white instead of Red, so can use red for error.
Used monitor-content to turn window label yellow when a zsh prompt appears
in the non-current window (NOTE: The matched string is pretty fuzzy)

Set bell colors to make any non-current window that has a bell be red, and
modified my zsh prompt to emit a bell on error. Fortunately, I don't
actually hear this bell, at least not while in tmux in gnome-term.

Yay! Been wanting this for a long time.