zsh: v() helper for launching vim. Pure zsh version
Opens given files in an existing vim within this shell and
foregrounds it. If there are none, launches a new vim. If there
are multiple, asks user which to send it to.

Replaces the python-based solution that ended up being overkill.

Getting the shell job name right requires writing the command to be
launched to a tmp file followed by &. Source the tmpfile to launch
it in the background, then immediately foreground it.