vim: augroup hcc_c to define some abbreviations for C
WIP - These aren't quite working like I want. I need a way to suppress
the space character that the user has typed after the abbreviation name,
as that is undesirable for many of these cases.

Also, if(, etc. might should be changed to a plain mapping, because it
doesn't work if I start typing a variable right after.

Maybe a better approach for thse is to use syntax highlighting to
make the lack of space red?