tmux: Create tmx script to manage cloned sessions
Use this as a wrapper for tmux. Give it the name of a session
(default "0" if none specified). If this session doesn't exist, it
will create it. If it does exist, create a clone with unique name
and set the destroy-unattached option. Thus, creation and cleanup
of clones is now automatic!