Add tmuxenv.sh: Function to set some env from the current tmux client
Typical use:
ssh -X <machine>
tmux attach <some_session>

At this point, $DISPLAY in this shell will be whatever it was when
the shell was first started. But we want it to tunnel back through
this ssh. Running tmuxenv will set $DISPLAY to the value inside the
current client, which was started within this ssh.

TODO: Handle SSH_AUTH_SOCK better. This typically unsets it when run
within an ssh session, which detaches this shell from its original
ssh agent. This is often not desired (although it could when using
ssh agent forwarding). So, may want to parameterize this behavior.