zsh: Add single-ssh-agent
Runs during initilization of interactive zsh. If it's not already running,
it launches an ssh-agent listening on a well-known socket path.
This makes ensures that every shell started by the same user always
connects to the same running instance of ssh-agent instead of sometimes
being connected to the GUI one launched by the gnome session, and
sometimes not being connected to an ssh-agent at all (e.g. in the case
of an ssh login).

This has the conesequence that no command line shell will share the
same agent as the GUI apps, so you might have to enter passphrases
up to twice.