vimrc-common: Use g instead of a as first letter of my tab mappings
Because using a caused a delay when trying to enter insert-append mode
and could cause a problem if the first character to be inserted was the
second character of one of the tab page mappings and the user didn't
delay long enough.

Most of the g+char mappings were free, but I did not want to override gn
(visually select the next search match, or extend existing visual selection)
so I mapped :tabnew to g<Enter> instead.