vimrc-common: Make <Space> be prefix of window commands, like <C-W>
Also, custom mapping <Space><C-J> and <Space><C-K> to move current
window up or down.  Need to refine this for left/right.

I really hate using <C-W> for window commands, not only because this is
inconvenient to type, but also because it means "close tab or window" in
most other applications, so I don't want to get into the habit of using
it and then breaking things if the wrong application has focus.