zsh: Set grep & less command line options
The GREP_OPTIONS env var has been loudly deprecated in Fedora 22, so
use an alias instead. As noted in the comment, this causes loss of
some functionality.

for less, want --quit-if-one-screen an --ignore-case to be default,
so use LESS env var to accomplish this (necessitates -R also in order
to see color output correctly)