hcc-vim.zsh: Fix v --split (vs) for existing vim servers
Bug fix: v --split is supposed to open each listed file in a new split
window. This worked fine if we were launching a new vim (in which case,
we merely added a -o to the vim command line). But, when communicating
with an existing vim, it was glitchy because the vim --remote +cmd
syntax did not execute the +cmd command until after loading the file.
The result was that the first listed file would end up occupying two
splits, and whatever was in the existing window would end up being
replaced by the first listed file.

Solution is to use vim --remote-send to send keystrokes that will
properly load each file into a split. Also added a couple of window
commands to cause them to load in order from top to bottom, cuz the
default reverse order was annoying!