Add bin/sshx: Wrapper for ssh which attaches xpra to remote
Adds an arbitrary local TCP port fowarding and launches a background
function to repeadedly attempt to xpra attach to it as long as it
gets "Connection refused" up to a max number of attempts.

The default remote Xpra port to attach to is 5000 + $XPRA_DISPLAY
number. Override this with --xpra-port NNN, which must be the first two
parameters, because the rest is passed to ssh verbatim.