tmux: Fix backward-compatibility with version < 2.1
The conditionals for handling mouse mode didn't work. Instead of testing
the tmux version, let's test for the existance of the options in
question. This is easier, since versions can be non-numbers, like 1.9a,
and cleaner since we don't have to embed the relevant version number in
our conf file.

Also make the key binding for toggling mouse mode work with both older
and newer versions of tmux.