hcc-vim.sh: Really fix v --split (vs, vts) for existing Vim's
The changes from Sept 1, 2015 made these commands work, as long as
either the shell's current directory and Vim's current directory
matched, or absolute paths are specified. This is because we were using
send-keys to send the filename to the :split command, so Vim had no idea
what the shell's current directory is.

Fix is to split the creation of the split and opening of the file into
two separate vim --remote... commands since it knows now to take your
shell's current directory into account when you simply use --remote to
directly open a file.

This commit also fixes the bug where vts (v --tab --split) would leave a
blank window behind in the new tab.