hcc-vim.zsh: Integrate with Xpra (--noxpra to override)
Use the xp function in xpra.zsh to always try to run vim with DISPLAY
and XAUTHORITY set for the default Xpra instance that I launch with tmxx
(tmux wrapper). This allows Vim to survive crash of surrounding X server
and to work correctly when launched from headless ssh session because
the Xpra server always stays running.

This will fail miserably if Xpra is not running or we can't talk to it.
Add --noxpra to the end of all commands: v, vs, vt, vts, vnew, vnews
to override this behavior and just use the shell's existing values for
$DISPLAY and $XAUTHORITY.  Useful for running quick vim when no tmux
session is running (although Xpra does not require tmux to work, I
typically launch it with my tmxx wrapper script).