sshx: Use timeout for retry, not count. Fix --sharing
We used to retry the xpra attach 40 times with a 1-second delay between
each. However, this caused weird behavior, because, after a successful
session, it would resume retrying a number of times. We have no way of
determining whether the xpra ever attached successfully, so we'll use
wall clock time as a heuristic.

In one funny episode, after a long session connected to work from my
laptop (which was still connected), I connected from my desktop. Because
--sharing was not enabled, this cased xpra on the laptop to terminate.
But its sshx script then retried, and so did the one on the desktop, so
that it rapidly connected and disconnected dozens of time, as the two
scripts "fought" with one another! This change will prevent that.

Also enable --sharing=yes so that both machines could attach to the
remote Xpra at once. This had been left out by mistake.