bin/pager: A wrapper script for less to fix --quit-if-one-screen
Use a wrapper script for less to fix the fact that --quit-if-one-screen
will, by default, eat output less than a screen completely (shows it on
the alternate screen but then exits, abandoning the output). The script
tee's less's input to a temp file, then, after less exits, if the file
would have fit on a screen, it cats it to the terminal.

This fix was linked to from here:
http://unix.stackexchange.com/questions/107315/less-quit-if-one-screen-without-no-init