tmux: Make multiple sessions much nicer with multiple clients
Added new script tmux-sessions, which aims to mostly "hide" the annoying
"clone" sessions that must be created in order for multiple clients to
share a session in a sane way (that is, where they aren't forced to be
viewing the same window all the time).

Unfortunately, this can cause clone sessions to build up after the
associated client has exited. These can be cleaned up using
"tmux-sessions clean". However, these sessions will typically be re-used
if a new client is created.