tmx: Play nicely with clone management in new tmux-sessions script
Never attach directly to a base session, but always create a clone, even
for the first client. This allows every client to have a "number" equal
to the number at the end of the clone sessions's name (e.g. "blah_C2").
When coming up the number for a new client, choose the lowest one for
which there does not already exist a client having that number
(regardless of whether clone sessions already exist having this suffix).
This way, we re-use existing clones if they were left around by a
now-exited client.

If invoked from within tmx, run tmux-sessions switch-to <base_name> to
switch the current client to an appropriately-named clone of the target
base session.

If the argument is "clean", run "tmux-sessions clean".  Just a quick
pass-through to save user a bit of typing.