tmux-session: Fix switch-* when must create new clone session
Tmux 1.9a errors out if you run tmux new-session -d ... to create a new
clone session with the message: "sessions should be nested with care,
unset $TMUX to force", even though the -d flag means we're not actually
nesting a session.  This was fixed by tmux 2.1.

But, for backward compatibility with tmux 1.9a, unset TMUX whenever we
have to create a clone session.