vim: Bugfix <Enter> not working in QuickFix window
The reason <Enter> did not jump to the location of the error under the
cursor is that I had re-mapped <Enter> to be ;  (that is, repeat the
last motion, since I prefer ; to be like :).

Fix it by mapping it to a function that preserves its original meaning
when executed within the current buffer is QuickFix.