Add xpra-sync-clip script to keep clipboard synchronized
Xpra's built-in clipboard synchronization is not reliable. It sometimes
causes "sync storms" which, when detected, cause xpra to cease
synchronizing clipboards, with no way to resume it without killing Xpra.

So, I'll use this script to use xclip to keep the Xpra and main X
session clipboards in sync instead.  It separately synchronizes the
PRIMARY and CLIPBOARD selections between the two X servers, but does not
synchronize these selections between each other.

Unfortunately, there is no way to make this event-driven from outside of
the X server, os it relies on expensive polling. So we have to keep the
polling interval low (currently 2 seconds for a full cycle).

It's best to disable Xpra's own clipboard sync when using this.

TODO:
* Make it convenient to start an instance of this to synchronize a
  remote Xpra clibpard tunneled through ssh with the local Xpra or main
  X session clipboard. Requires parameterization.

* Change my xpra start-up command not to enable clipboard sync

* Automatically start this whenever I start my normal local Xpra,
  i.e. lxpra()

* Automatically start a remote-sync instance when connecting via sshx.

* (maybe) Add some logging of clipboard changes?