.tmux/hcc-keys.conf: Remove mouse on/off key binding
Commit 38f3ad2 ("tmux: Change mouse toggle to Ctrl-M, Ctrl-m") is broken
for two reasons:
1. C-m is the same as "Enter", so typing <prefix><Enter> was disabling
the mouse, which annoys since I often use "Enter" to mean "Cancel the
prefix state".

2. The binding to turn the mouse back on did not work because the tty
cannot distinguish C-m from C-M (that is, Ctrl+Shift+m).

I still don't want to conflict with the default use of m and M to mark /
unmark a pane. And Using Alt-m doesn't work because gnome-terminal eats
it. So, rather than fix this, I'm just going to drop the mouse on/off
bindings altogether. I never use them. And, on the rare occasion when it
may be warranted, it's easy enough just to type :set mouse on.

Fixes: 38f3ad2 ("tmux: Change mouse toggle to Ctrl-M, Ctrl-m")