tmux: Remove root binding of Alt-PgUp/Down for nested tmux
I had "root" (i.e. non-prefix) bindings of Alt-PgUp/Down to select the
window to the left or right of the current one, but I never actually use
this binding.

Meanwhile, this masked the ability to use <prefix> Alt-PgUp/Down binding
within a nested session, which swaps the current window with the one to
the left or right.

So get rid of the unused binding to make nested tmux easier.