Prevent Xpra client from scaling windows
By default, if the client host has a very large display pixel count (as my
3-monitor setup has), it will assume that we're HiDPI and will scale up the
windows from the Xpra server, which is very annoying. It even does this when
the server is on the same machine, which just makes no sense!

Disable this in my xpra client command line, both for remote clients
(bin/sshx), and for the local client (xpra.zsh:xprattach()).