xpra.zsh: xprastart: Workaround Xpra bug daemonizing before ready
As of v1.0.1, Xpra has a bug where it daemonizes before it is ready to
receive client connections (by around 1.5 seconds on my machine!).

So make the xprastart() function poll the TCP port until it is open
before returning.

Needed because lxpra() will immediately create a local client connection
as soon as the server is started.