vim: Re-enable wrapping of normal text lines
This was turned off because I was getting some wrapping when editing code,
which is obviously annoying. However, disabling this broke wrapping in cases
when it is needed (such as writing git comments). Since re-enabling it, I do
not see code wrapping either, so I think I just had something misconfigured
that day.