vim: Increase keystroke timeoutlen
I decreased this to only 100ms awhile back because it was taking a really long
time to actually exit insert mode when Esc was pressed. Although this short
timeout is OK for keys that transmit multiple bytes (such as arrows or
function keys), it makes bindings like leader+<anything> hard to type.

This commit compromises with a bit longer of a timeout.