tmux: Change mapping to select windows 10-19 to use prefix
I previously mapped M-0 through M-9 (i.e. Alt-0 through Alt-9) witout
the prefix (-n mappings) to select windows 10-19, and the same with prefix to
select windows 20-29.

But this "breaks" nested tmux sessions, because lack of prefix means that the
outer tmux always interprets the command - no way to direct it to the inner
tmux.

So, we'll use prefix + M-0 through M-9 to select 10-19 and just have no
mapping for selecting 20-29. I have yet to encounter a need for that many tmux
windows, so can live without it.