tmux: Remove unbinds
When I first started using tmux, it was illegal to bind a key if it was
already bound, necessitating unbinding each key before binding it in the
config file. Otherwise, you could never "re-source" the config file within a
tmux instance.

However, the unbinds became unnecessary sometime prior to tmux 2.1, so
removing the cruft.