zsh: setprompt: Show number of background jobs and vim flag
If this shell has background numbs, show the count in purple. If one of them
is an instance of vim, show a vim flag since I especially care which shells
contain a vim session. If there are no other background jobs besides vim(s),
suppress the count display since it is redundant (although I suppose it isn't
if there are multiple vim's. Oh well.)