hcc-zle.zsh: Clean up bindings and fix file-only completion
Move all bindings that were still in .zshrc to .zsh/zsh.d/hcc-zle.zsh,
removing several out-dated ones in the process.

This change also re-implements Shift-Tab to force filename-only completion
because the existing implementation was mostly broken (didn't work if the path
contained a path separator / or a wildcard, etc).