vim: Reduces timeoutlen cuz Neovim makes you wait after Esc
With Neovim, I have to wait the duration of timeoutlen after pressing Esc for
it to actually exit Insert mode (wasn't like that on Vim 7.4). So, decrease it
to a tolerable amount without making me have to rush mappings too much.

Ultimately, this is a bit of a poor compromise, and something better is
desired.