zsh: Make v, vs, vts, use Neovim when available
The versions that use Vim are renamed to oldv, oldvs, oldvts, etc.
If nvim is not installed, the Vim versions remain.

The implementation uses nvr "Neovim Remote" command, which is much cleaner
than relying on the X server for vim --remote. This alleviates my need to
always be running Xpra in the background.