vim: LanguageClient: Fix focus stealing by pop-up
The hover action shows info about the symbol under cursor in a pop-up
window (Nvim 4) or a split window (earlier versions). But the way I had
the hover event triggered from events could cause two of them to occur
in a row, which the LanguageClient plugin interprets as a command to
move the focus to the pop-up window, stealing focus from the window I am
working in, which was frustrating.

So, revamp the way events are handled to only call the hover action when
it makes sense to do so.