install: Re-add .vimrc to for compatibility with regular Vim
The Neovim changes broke regular Vim by no longer installing .vimrc. Add
this back and make it a symlink instead of a regular file. Long ago,
this had been changed to regular file copied from config/vimrc-template
so that it could easily be edited with machine-specific changes. For
better consistency with .vimrc, etc, make it souce ~/.vimrc-local if it
exists instead.