Add packages list from Ubuntu 18.04 (light)
These lists will serve as a default set of preferred packages for
me to install on a new machine. I have them broken into command-line
packages, GUI packages, and special packages that require a PPA or
prefer to install as a snap or prefer to not typically install but
just want them as a reference.