Fix 'less' compatibility (bin/pager)
The workaround script bin/pager is no longer needed on Ubuntu 20.04, and
it caused duplicate output of short files using less. So put a version
check to just invoke less normally in that case.

Also remove --quit-if-one-screen from LESS env var by default since that
option doesn't actually work on earlier less versions, and the script
worked around it anyway. It could cause trouble in the rare case that
the script and less disagreed about whether the content could fit on one
screen.