vim: Switch from coc-python to coc-jedi
Necessary because coc-python is discontinued by its author, who
recommends people use coc-jedi instead. Note that I am not able to get
coc-jedi to work with Python 2.7 despite removal of support for that not
being documented AFAICT. So, it looks like I have reaced the end of the
road fo a good Vim experience with Python 2.7. Can't wait to leave it
behind fully.