nvim: Make <Tab> work for to accept completion
But still allow it to indent when not doing completion. This is
surprisingly hard. It doesn't quite do what I want because I can't make
it select next item starting with the FIRST one. So instead, I just make
it directly confirm the first one. Use arrow keys to actually cycle
through stuff.