zsh: setprompt: Add username. Add $prompt_flags to configure
Add optional username@ in front of hostname in prompt. Turn it red if
superuser.

Instead of individual flag files to control optional components of the
prompt, introduce prompt_flags string. Each character present in the
string enables an optional component:

u: username@
s: "SUDO" flag
g: git info

This is done to make it easy to disable the SUDO flag and git info when
zsh is running from within a snap.

All flags are enabled by default (i.e. if the variable is unset when
setprompt is run by .zshrc)