.zshrc: Clear out $TMUX if tmux server not reachable.
If, for some reason, $TMUX is set but there is not actually a reachable
tmux server, then you can't start tmux. And ranger won't run.
This can happen if you enter an environment with a different VFS
namespace from a shell running in tmux (such as snap run --shell..., or
possibly in a chroot (although the sudo used to get into chroot tends to
clear env vars).