nvim: Add NVIM_BASIC_CONFIG mode (env var)
When enabled by setting this env var to 1, specify a much-simplified
set of plugins. To be used in embedded environments such as inside a
snap. Lacks language servers, completions, and uses my old 256-color
scheme "hcc" instead of onedarkpro.

Also configure Packer to store packer_compiled.lua under
~/.local/share/nvim instead of under ~/.config/nvim because the latter
location can be read-only when running from a snap. Also, since this is
a generated file, it makes sense to keep it in the data area rather with
user-specified config files, which are typically under revision control.