install: Make most of it optional
Add options --no-apt, no-locale, --no-nodejs, --no-nvim, --no-tmux
that allow you to skip those pieces of the install. Also --minimal is
short for all of the above, so the only thing that is done is
creating/updating symlinks to the source files from home account
dotfiles.

This is useful when the environment may already include some of these
things. For example, Badger's robot Yocto image doesn't have APT but
includes in the image all of the packages that this script gets from
APT. Likewise, it installs locales, Node.js, and Nvim from separate
bitbake recipes.