nvim: wildmenu: make <CR> advance to next subtree
When using the up/down arrow keys to navigate the vertical wildmenu, the
right arrow key advances to the next subtree. When using tab/shift-tab
to navigate, however, there is no good way to do this. Pressing enter
simply results in nvim opening the currently selected file/directory.
Instead, when in a wildmenu, remap <CR> to <Down> so that pressing enter
advances to the next subtree. Once the wildmenu is closed, pressing
enter will open the chosen file/directory.