nvim: Add better-escape plugin
Eliminates the annoying delay of nvim's timeoutlen when typing a j or k
in insert mode due to the mapping of jk to <Esc>.

Unfortunately, the better-escape plugin only works in Insert mode,
so we continue to rely on a standard mapping for Command and
Operator-pending modes and to not be able to have this mapping at all in
Visual mdoe because the delay is intolerable in that mode. It also
forces me to continue to keep timeoutlen quite short (200ms) while I
prefer to lengthen it to 500 or more.

Maybe eventually, I'll rewrite better-escape to work in more modes.
However, a different approach must be taken that doesn't depend on the
InsertCharPre autogroup event.