nvim: cmp: Fix <Esc> and <Down> mappings
New config mapped <Esc> to just cancel the completion, rather than exit
back to Normal mode, which is great if you've already started scrolling
through the selections. But if you're just typing along when
completions have popped up spontaneously, the user still expects <Esc>
to go directly to Normal mode. So fix this mapping accordingly.

New config broke the <Down> mapping because it fails to call the
fallback() function when completion window is not open. As a result, it
became impossible to move the cursor down while in Insert mode.