nvim: Don't let x clobber the "cut" register
Normally, the normal mode 'x' command (delete character) writes to all
of the following registers:

"  The "cut" register
*  The gui clipboard register
+  The gui selection register
-  The "small deletion" register

Use a mapping to make it only write to the '-' register from now on so
that it no longer clobbers the last yank or deletion. The only reason
for making it do this rather than write to the black hole '_' register
is that you might actually 'x' can actually delete many characters if
given a count.