nvim: Add nvim-neoclip
Provides clipboard history accessible from Telescope.
Mapping:
<leader>s'   - Open telescope with past cut/yanks
<leader>s2   - Open telescope with past macro recordings

In telescope, select an entry with <cr> to put it into the ", *, +
registers (for clipboard history) or into the q register (macro
history).

Or use p or P in normal mode to paste the item directly without
modifying any registers, or q to directly replay the macro.