nvim: Fix <Down> not working in cmd mode
It turns out that the cmd mode mappings in nvim-cmp were interfering
with the ability of <Down> to function in command mode (including in the
pop-up wildmenu and just to move down in the command history).

Remove nvim-cmp's command-mode mappings, since we do not use nvim-cmp in
command mode anyway.

I did experiment with enabling cmp for command mode, and although this
allows some nice fuzzy finding, it is otherwise a bit glitchy. The
native pop-up menu is good enough, IMO.