zsh: Remove depdency lsb_release; Cleanup less config
Remove our dependency on lsb_release being installed in both the install
script and in .zshrc.

The install script will now assume we're on a modern Ubuntu if it can't
run lsb_release (all it really wanted to know is whether we're on xenial
or not anyway).

Cleanup: Remove the bin/pager, which was a workaround for less's
--quit-if-one-screen option being broken in old versions of less. Now,
just include that option in export LESS and assume it will work.

The bin/pager script was already just using the option directly if it
detected that we had a modern LESS installed, but since .zshrc only
configured the PAGER to be bin/pager if we're on Ubuntu (which it
detected using lsb_release), it meant that we didn't get the
--quit-if-one-screen behavior on non-Ubuntu systems; now we do.