nvim: Make headless installation of all plugins work
The ability to synchronously install all plugins and then exit vim
headlessly was lost quite some time ago. Write some lua code to get it
back and make our install script invoke it. Note, it is still sometimes
racy because there is no way to completely sure that treesitter is
finished compiling all parsers, but it mostly works for now.

Also add --no-nvim-plugins option to install so that you can choose to
install nvim without its plugins. Make it possibly to use --no-nvim but
not --no-nvim-plugins to ONLY install the plugins but not Nvim itself.
Useful if the OS already has a recent version of Nvim installed.