install: Install rust for user via rust-install-script
Some nvim plugins need cargo and rustc to install rust programs. An
attempt was made to just use cargo from but the version of rustc bundled
with Ubuntu 22.04 is not new enough to build some nvim plugins, so that
didn't work. This method will install the latest rustc just for the
current user.