zsh: Always set TERM to xterm-256color when not in tmux
This is because the default value of TERM is often the 8-bit-color-only
"xterm", so colors look terrible. This is often the case when starting a
shell in a docker container (unlike when using ssh, TERM is not
inherited from the host's shell in docker). The default value of
TERM is also 8-bit only when logging directly into the console character
terminal, yet it is actually capable of at least 256 color.