Cleanup - move lists to notes/ and scripts to bin/
Non-functional. Just cleaning up the root directory of this repo a bit.
Move the various package lists that I made for reference under a new notes/
subdirectory. NOTE: These package lists are currently a bit out of
date.

Also move a couple of scripts from the root directory to the bin
directory.