install: Detangle tmux plugins from hcc-dotfiles workspace
We used to create a top-level symlink from ~/.tmux ->
<hcc-dotfiles>/config/.tmux, with config/.tmux/plugins/tpm pre-existing
there as a submodule of this repo.

Now, instead, make ~/.tmux be a real directory, and create a symlink
therein pointing to each individual file in
<hcc-dotfiles>/config/.tmux/*.

Remove tpm as a submodle of hcc-dotfiles, and instead have the install
script clone it at ~/.tpm/plugins/tpm. Then TPM itself will clone its
set of configured plugins also in ~/.tpm/plugins.

Why? To separate things that are installed by this configuration (i.e.
the tmux plugins installed by tpm) from the hcc-dotfiles workspace,
which makes the packaging of a binary realease cleaner.