release: Add fixup-username-after-tarball-extract
Because the release process builds the tarballs as user `dotbuilder`
within Docker, we found a few cases where files in the tarball end up
referring to hard-coded paths containing `home/dotbuilder`.

Add a new script, `fixup-username-after-tarball-extract` that the user
can run after extracting tarballs to fixup known references to the
user's own username.

Also, limit the problem by excluding pre-compiled Python bytecode files
(pyc) from the tarballs to begin with. They all contain the `dotbuilder`
username. This also shrinks the tarball slightly with no penalty since
the .pyc files will be built when run for the first time anyway.

LIMITATION: A couple of Rust-compiled executables contain the username
`dotbuilder`, which the script does not fixup. See comment in the
script.